MEX-962: map single batch to an array to keep consistency

+++ src/modules/dust-converter/models/dust.converter.model.ts
@@ -49,7 +49,7 @@ export class DustConvertPathModel {
     @Field(() => Float)
     amountOutShort: number;
 
-    @Field(() => Int)
+    @Field(() => Float)
     splitPpm: number;
 
     @Field(() => [DustConvertSwapModel])
